+++ packages/media/src/icon/SuccessIcon.tsx
@@ -1,13 +1,13 @@
 import React from 'react'
 
-const AttentionIcon: React.FunctionComponent = () => (
-  <svg viewBox="0 0 40 40" focusable="false">
-    <circle cx="20" cy="20" r="20" />
+const SuccessIcon: React.FunctionComponent = () => (
+  <svg viewBox="0 0 24 24" focusable="false">
+    <circle cx="12" cy="12" r="12" />
     <path
       fill="#FFF"
-      d="M16.94 26.06a1.5 1.5 0 0 0 2.12 0L29 15.122A1.5 1.5 0 0 0 26.879 13L18 23l-2.94-3.06a1.5 1.5 0 0 0-2.12 2.12l4 4z"
+      d="M8.707 12.293a1 1 0 1 0-1.414 1.414l2.5 2.5a1 1 0 0 0 1.414 0l5.5-7.5a1 1 0 1 0-1.414-1.414L10.5 14.086l-1.793-1.793z"
     />
   </svg>
 )
 
-export default AttentionIcon
+export default SuccessIcon