+++ packages/mobile/src/typography/Title.tsx
@@ -1,6 +1,6 @@
 import React, {FunctionComponent} from 'react'
 
-import {Typo} from '@qiwi/pijma-core'
+import {Breaker, Typo} from '@qiwi/pijma-core'
 
 export interface TitleProps {
   tag?: 'h1' | 'h2' | 'h3' | 'h4' | 'h5' | 'h6'
@@ -41,7 +41,7 @@ export const Title: FunctionComponent<TitleProps> = ({tag, size, color = 'defaul
     height={TitleHeight[size]}
     weight={TitleWeight[size]}
     color={TitleColor[color]}
-    children={children}
+    children={<Breaker children={children}/>}
   />
 )