+++ packages/mobile/CHANGELOG.md
@@ -1,3 +1,17 @@
+# [@qiwi/pijma-mobile-v1.40.0](https://github.com/qiwi/pijma/compare/v1.39.0...v1.40.0) (2019-05-21)
+
+
+### Features
+
+* breaker ([10bc3c0](https://github.com/qiwi/pijma/commit/10bc3c0))
+
+# [@qiwi/pijma-mobile-v1.39.0](https://github.com/qiwi/pijma/compare/v1.38.0...v1.39.0) (2019-05-13)
+
+
+### Features
+
+* drop up ([58e91ad](https://github.com/qiwi/pijma/commit/58e91ad))
+
 # [@qiwi/pijma-mobile-v1.37.0](https://github.com/qiwi/pijma/compare/v1.36.1...v1.37.0) (2019-04-10)